MAINT: Fix duplicate symbols
