+++ CMakeLists.txt
@@ -17,6 +17,7 @@ add_library(libnmssearch STATIC
     src/command_parser.cc 
     src/build_index.cc 
     src/dump_index.cc
+    src/kmer_space.cc
     src/nw_align.cc 
     src/sequence_reader.cc 
     src/query_index.cc
