+++ src/kmer_space.h
@@ -27,17 +27,4 @@ protected:
     DISABLE_COPY_AND_ASSIGN(KMerSpace);
 };
 
-// Implementation
-
-std::string KMerSpace::StrDesc() const
-{
-    return "KMer cosine distance";
-}
-
-float KMerSpace::HiddenDistance(const similarity::Object *obj1, const similarity::Object *obj2) const
-{
-    auto s1 = (*m_spectra)[obj1->id()], s2 = (*m_spectra)[obj2->id()];
-    return cosine_distance(s1, s2);
-}
-
 #endif // _SPACE_H_
